Migrate to datajoint v2  (#74)
* AI roi masks fix

* AI improve ROI canvas

* Fix correlation calculation and ROI mask shifting in compute_corr_map_match_indexes

* AI fix corr_roi_mask

* add verbose for check_filesystem_consistency

* undo change to _compute_traces

* fix correlation masks

* improve roi mask scanning

* improve color noise processing

* fix default frame rate to 60 (not 1/60)

* improve FastSTA

* update trigger upsampling in noise preprocessing

* fix QDSpy log reader

* add PresentationLogTemplate

* improve log matching

* add time matching for stim logs

* fix

* fix preslogs defintion

* improve STA and cascade tables

* use numpy files for ROI masks

* improve type hints

* allow ROI mask scanning from different sources

* Github workflow, djimaging integration test using mysql 8

* Better tests

* local testing instructions, review feedback

* Test against py3.12, and djimaging 2.0 (should fail)

* Run again against dj 0.14, test python 3.12 until 3.14 in normal tests

* Datajoint v2 version

* get rid of migration script

* Suggestions from codex review

* djconfig template update

* Use load_array consistently

* Only one reference folder, use file paths against that reference folder

* Use blob for smaller data sizes

* Deps updates

* readd concurrency for traces in example notebook

* Import Classifier tables

* More blobs

* Datajoint interface changes for celltype assignment

* Use classifier for scikit-learn v1.7

* Minimum version requirements for some packages (e.g. numpy)

* Tested autorois: fixed restricting on string file path bug and opening dj file object direclty with pd.read_csv

* fixed a minor bug that tried to store a tupple in a field that requires numpy arrays.

* Fix typo: bic -> aic

* Fixed FastSta restriction bug: self.params_table.stimulus_table has all stimuli and if restrictions is not set then it throws an error in fetch1. restricting on params_table leads to only correct stimuli to be used.

* Fix test with update in testing code

* Remove deprecated/v1 and duplicate code

* A bit more cleanup

---------

+++ djimaging/tables/clustering/external_clustering.py
@@ -21,10 +21,10 @@ class RoiClusterTemplate(dj.Manual):
     @property
     def definition(self):
         definition = """
-        clustering_id : tinyint unsigned
+        clustering_id : int32
         -> self.roi_table
         ---
-        cluster_idx : int
+        cluster_idx : int32
         """
         return definition
